Split out Results, use simple mui imports
