+++ src/workbench/search/Search.tsx
@@ -1,21 +1,19 @@
 
-import React from 'react';
+import React, { useState } from 'react';
 
-import { Link, Paper } from '@mui/material';
+import Button from '@mui/material/Button';
+import Snackbar, { SnackbarCloseReason } from '@mui/material/Snackbar';
 
 import { useProgressStateStore } from '../state/ProgressState';
-
-import {
-    Table, TableBody, TableCell, TableContainer, TableHead, TableRow,
-} from '@mui/material';
-
 import { useSocket } from '../socket/socket';
 import { useWorkbenchStateStore } from '../state/WorkbenchState';
 import { useSearchStateStore } from '../state/SearchState';
 import {
-    Row, getGraphEmbeddings, addRowLabels, addRowDefinitions,
+    getGraphEmbeddings, addRowLabels, addRowDefinitions,
     addRowEmbeddings, computeCosineSimilarity, sortSimilarity,
 } from '../state/row';
+
+import Results from './Results';
 
 import SearchInput from './SearchInput';
 
@@ -28,25 +26,21 @@ const Search : React.FC <SearchProps> = ({
     const addActivity = useProgressStateStore(
         (state) => state.addActivity
     );
+
     const removeActivity = useProgressStateStore(
         (state) => state.removeActivity
     );
 
+    const [error, setError] = useState<string>("");
+
     const socket = useSocket();
 
-    const setSelected = useWorkbenchStateStore((state) => state.setSelected);
-    const setTool = useWorkbenchStateStore((state) => state.setTool);
     const setEntities = useWorkbenchStateStore((state) => state.setEntities);
 
     const view = useSearchStateStore((state) => state.rows);
     const setView = useSearchStateStore((state) => state.setRows);
 
     const search = useSearchStateStore((state) => state.input);
-
-    const select = (row : Row) => {
-        setSelected({ uri: row.uri, label: row.label ? row.label : "n/a" });
-        setTool("entity");
-    }
 
     const submit : React.FormEventHandler<HTMLFormElement> = (e) => {
 
@@ -81,10 +75,13 @@ const Search : React.FC <SearchProps> = ({
 
                 removeActivity(searchAct);
 
+                setError("");
+
             }
         ).catch(
             (err) => {
                 console.log("Error: ", err);
+                setError(err.toString());
                 removeActivity(searchAct);
             }
         );
@@ -92,58 +89,41 @@ const Search : React.FC <SearchProps> = ({
         e.preventDefault();
     
     }
+  const handleClose = (
+      _event : React.SyntheticEvent | Event,
+      reason?: SnackbarCloseReason,
+  ) => {
+      if (reason === 'clickaway') {
+          return;
+      }
+      setError("");
+  };
+
+  const action = (
+      <>
+          <Button color="error" size="small" onClick={() => setError("")}>
+              Close
+          </Button>
+      </>
+  );
 
     return (
         <>
 
             <SearchInput submit={submit}/>
 
+            <Snackbar
+                open={error != ""}
+                autoHideDuration={6000}
+                onClose={handleClose}
+                message={error}
+                action={action}
+                anchorOrigin={{ vertical: "bottom", horizontal: "center" }}
+            />
+
             {
                 view.length > 0 &&
-                    <TableContainer component={Paper} sx={{ mt: 4 }}>
-                      <Table sx={{ minWidth: 450 }}
-                          aria-label="table of entities"
-                      >
-                        <TableHead>
-                          <TableRow>
-                            <TableCell>Entity</TableCell>
-                            <TableCell>Description</TableCell>
-                            <TableCell>Similarity</TableCell>
-                          </TableRow>
-                        </TableHead>
-                        <TableBody>
-                            {
-                                view.map((row : Row) => (
-                                    <TableRow
-                                      key={row.uri}
-                                      sx={{
-                                          '&:last-child td': { border: 0 },
-                                          '&:last-child th': { border: 0 }
-                                      }}
-                                    >
-                                        <TableCell component="th" scope="row">
-                                          <Link
-                                              align="left"
-                                              component="button"
-                                              onClick={
-                                                  () => select(row)
-                                              }
-                                          >
-                                              {row.label}
-                                          </Link>
-                                        </TableCell>
-                                        <TableCell>
-                                            {row.description}
-                                        </TableCell>
-                                        <TableCell>
-                                            {row.similarity!.toFixed(2)}
-                                        </TableCell>
-                                    </TableRow>
-                                ))
-                            }
-                        </TableBody>
-                      </Table>
-                    </TableContainer>                
+                <Results/>
             }
 
         </>
